requirements.txt
ultralytics
Pillow
os
boto3
io